Energy calculation fix
The energy calculation relies on correct readings of the last 2 states
of the power attribute. The problem was that if the event is sent
without state change = true, and the power value is the same as the
previous event, the state is not saved and the statesSince() method used
for the energy calculation will return the last 2 states when the power
changed, which is not what we want.