Fix energy calculation and remove from activity feed
The change that was checking if(energyValue) broke the energy
accumulation after reset because then the value is 0.0 and this code
line will prevent it from ever increasing. The correct check should be
if(energyValue != null).
Also I added displayed: false to all events to remove the messages from
the activity log because other users have complained that it is too
"chatty".